+++ test/unit/EthereumVaultConnector/Call.t.sol
@@ -11,7 +11,7 @@ contract EthereumVaultConnectorHandler is EthereumVaultConnectorHarness {
     function handlerCall(
         address targetContract,
         address onBehalfOfAccount,
-        uint value,
+        uint256 value,
         bytes calldata data
     ) public payable returns (bytes memory result) {
         result = super.call(targetContract, onBehalfOfAccount, value, data);
@@ -25,10 +25,7 @@ contract CallTest is Test {
     EthereumVaultConnectorHandler internal evc;
 
     event CallWithContext(
-        address indexed caller,
-        address indexed targetContract,
-        address indexed onBehalfOfAccount,
-        bytes4 selector
+        address indexed caller, address indexed targetContract, address indexed onBehalfOfAccount, bytes4 selector
     );
 
     function setUp() public {
@@ -54,16 +51,12 @@ contract CallTest is Test {
         address nestedTargetContract = address(new TargetWithNesting());
         address controller = address(new Vault(evc));
         vm.assume(
-            targetContract != alice &&
-                targetContract != address(evc) &&
-                !evc.haveCommonOwner(targetContract, alice) &&
-                !evc.haveCommonOwner(targetContract, account)
+            targetContract != alice && targetContract != address(evc) && !evc.haveCommonOwner(targetContract, alice)
+                && !evc.haveCommonOwner(targetContract, account)
         );
         vm.assume(
-            nestedTargetContract != alice &&
-                nestedTargetContract != address(evc) &&
-                !evc.haveCommonOwner(nestedTargetContract, alice) &&
-                !evc.haveCommonOwner(nestedTargetContract, account)
+            nestedTargetContract != alice && nestedTargetContract != address(evc)
+                && !evc.haveCommonOwner(nestedTargetContract, alice) && !evc.haveCommonOwner(nestedTargetContract, account)
         );
 
         vm.prank(alice);
@@ -72,30 +65,15 @@ contract CallTest is Test {
         Vault(controller).reset();
 
         bytes memory data = abi.encodeWithSelector(
-            Target(targetContract).callTest.selector,
-            address(evc),
-            address(evc),
-            seed,
-            account,
-            seed % 2 == 0
+            Target(targetContract).callTest.selector, address(evc), address(evc), seed, account, seed % 2 == 0
         );
 
         vm.deal(alice, seed);
         vm.expectEmit(true, true, true, true, address(evc));
-        emit CallWithContext(
-            alice,
-            targetContract,
-            account,
-            Target.callTest.selector
-        );
+        emit CallWithContext(alice, targetContract, account, Target.callTest.selector);
         vm.prank(alice);
-        bytes memory result = evc.handlerCall{value: seed}(
-            targetContract,
-            account,
-            seed,
-            data
-        );
-        assertEq(abi.decode(result, (uint)), seed);
+        bytes memory result = evc.handlerCall{value: seed}(targetContract, account, seed, data);
+        assertEq(abi.decode(result, (uint256)), seed);
 
         evc.reset();
         Vault(controller).reset();
@@ -113,27 +91,12 @@ contract CallTest is Test {
 
         vm.deal(alice, seed);
         vm.expectEmit(true, true, true, true, address(evc));
-        emit CallWithContext(
-            alice,
-            nestedTargetContract,
-            account,
-            TargetWithNesting.nestedCallTest.selector
-        );
+        emit CallWithContext(alice, nestedTargetContract, account, TargetWithNesting.nestedCallTest.selector);
         vm.expectEmit(true, true, true, true, address(evc));
-        emit CallWithContext(
-            nestedTargetContract,
-            targetContract,
-            nestedTargetContract,
-            Target.callTest.selector
-        );
+        emit CallWithContext(nestedTargetContract, targetContract, nestedTargetContract, Target.callTest.selector);
         vm.prank(alice);
-        result = evc.handlerCall{value: seed}(
-            nestedTargetContract,
-            account,
-            seed,
-            data
-        );
-        assertEq(abi.decode(result, (uint)), seed);
+        result = evc.handlerCall{value: seed}(nestedTargetContract, account, seed, data);
+        assertEq(abi.decode(result, (uint256)), seed);
     }
 
     function test_RevertIfDepthExceeded_Call(address alice) external {
@@ -146,11 +109,7 @@ contract CallTest is Test {
         evc.call(address(0), alice, 0, "");
     }
 
-    function test_RevertIfNotOwnerOrOperator_Call(
-        address alice,
-        address bob,
-        uint seed
-    ) public {
+    function test_RevertIfNotOwnerOrOperator_Call(address alice, address bob, uint256 seed) public {
         vm.assume(alice != address(0) && alice != address(evc));
         vm.assume(!evc.haveCommonOwner(alice, bob));
         vm.assume(bob != address(0));
@@ -159,12 +118,7 @@ contract CallTest is Test {
         vm.assume(targetContract != alice && targetContract != address(evc));
 
         bytes memory data = abi.encodeWithSelector(
-            Target(targetContract).callTest.selector,
-            address(evc),
-            address(evc),
-            seed,
-            alice,
-            false
+            Target(targetContract).callTest.selector, address(evc), address(evc), seed, alice, false
         );
 
         vm.deal(alice, seed);
@@ -173,10 +127,7 @@ contract CallTest is Test {
         evc.call{value: seed}(targetContract, bob, seed, data);
     }
 
-    function test_RevertIfChecksReentrancy_Call(
-        address alice,
-        uint seed
-    ) public {
+    function test_RevertIfChecksReentrancy_Call(address alice, uint256 seed) public {
         vm.assume(alice != address(evc));
 
         address targetContract = address(new Target());
@@ -185,12 +136,7 @@ contract CallTest is Test {
         evc.setChecksLock(true);
 
         bytes memory data = abi.encodeWithSelector(
-            Target(targetContract).callTest.selector,
-            address(evc),
-            address(evc),
-            seed,
-            alice,
-            false
+            Target(targetContract).callTest.selector, address(evc), address(evc), seed, alice, false
         );
 
         vm.deal(alice, seed);
@@ -199,10 +145,7 @@ contract CallTest is Test {
         evc.call{value: seed}(targetContract, alice, seed, data);
     }
 
-    function test_RevertIfImpersonateReentrancy_Call(
-        address alice,
-        uint seed
-    ) public {
+    function test_RevertIfImpersonateReentrancy_Call(address alice, uint256 seed) public {
         vm.assume(alice != address(evc));
 
         address targetContract = address(new Target());
@@ -211,12 +154,7 @@ contract CallTest is Test {
         evc.setImpersonateLock(true);
 
         bytes memory data = abi.encodeWithSelector(
-            Target(targetContract).callTest.selector,
-            address(evc),
-            address(evc),
-            seed,
-            alice,
-            false
+            Target(targetContract).callTest.selector, address(evc), address(evc), seed, alice, false
         );
 
         vm.deal(alice, seed);
@@ -225,10 +163,7 @@ contract CallTest is Test {
         evc.call{value: seed}(targetContract, alice, seed, data);
     }
 
-    function test_RevertIfTargetContractInvalid_Call(
-        address alice,
-        uint seed
-    ) public {
+    function test_RevertIfTargetContractInvalid_Call(address alice, uint256 seed) public {
         // call setUp() explicitly for Dilligence Fuzzing tool to pass
         setUp();
 
@@ -238,12 +173,7 @@ contract CallTest is Test {
         // target contract is the EVC
         address targetContract = address(evc);
         bytes memory data = abi.encodeWithSelector(
-            Target(targetContract).callTest.selector,
-            address(evc),
-            targetContract,
-            seed,
-            alice,
-            false
+            Target(targetContract).callTest.selector, address(evc), targetContract, seed, alice, false
         );
 
         vm.deal(alice, seed);
@@ -254,12 +184,7 @@ contract CallTest is Test {
         // target contract is the msg.sender
         targetContract = address(this);
         data = abi.encodeWithSelector(
-            Target(targetContract).callTest.selector,
-            address(evc),
-            address(evc),
-            seed,
-            address(this),
-            false
+            Target(targetContract).callTest.selector, address(evc), address(evc), seed, address(this), false
         );
 
         vm.deal(address(this), seed);
@@ -269,12 +194,7 @@ contract CallTest is Test {
         // target contract is the ERC1820 registry
         targetContract = 0x1820a4B7618BdE71Dce8cdc73aAB6C95905faD24;
         data = abi.encodeWithSelector(
-            Target(targetContract).callTest.selector,
-            address(evc),
-            address(evc),
-            seed,
-            alice,
-            false
+            Target(targetContract).callTest.selector, address(evc), address(evc), seed, alice, false
         );
 
         vm.deal(alice, seed);
@@ -283,10 +203,7 @@ contract CallTest is Test {
         evc.call{value: seed}(targetContract, alice, seed, data);
     }
 
-    function test_RevertIfValueExceedsBalance_Call(
-        address alice,
-        uint128 seed
-    ) public {
+    function test_RevertIfValueExceedsBalance_Call(address alice, uint128 seed) public {
         vm.assume(alice != address(0) && alice != address(evc));
         vm.assume(seed > 0);
 
@@ -294,12 +211,7 @@ contract CallTest is Test {
         vm.assume(targetContract != alice && targetContract != address(evc));
 
         bytes memory data = abi.encodeWithSelector(
-            Target(targetContract).callTest.selector,
-            address(evc),
-            address(evc),
-            seed,
-            alice,
-            false
+            Target(targetContract).callTest.selector, address(evc), address(evc), seed, alice, false
         );
 
         // reverts if value exceeds balance
@@ -313,9 +225,7 @@ contract CallTest is Test {
         evc.call{value: seed}(targetContract, alice, seed, data);
     }
 
-    function test_RevertIfInternalCallIsUnsuccessful_Call(
-        address alice
-    ) public {
+    function test_RevertIfInternalCallIsUnsuccessful_Call(address alice) public {
         // call setUp() explicitly for Dilligence Fuzzing tool to pass
         setUp();
 
@@ -325,9 +235,7 @@ contract CallTest is Test {
         address targetContract = address(new Target());
         vm.assume(targetContract != alice && targetContract != address(evc));
 
-        bytes memory data = abi.encodeWithSelector(
-            Target(targetContract).revertEmptyTest.selector
-        );
+        bytes memory data = abi.encodeWithSelector(Target(targetContract).revertEmptyTest.selector);
 
         vm.prank(alice);
         vm.expectRevert(Errors.EVC_EmptyError.selector);
