+++ test/unit/EthereumVaultConnector/Callback.t.sol
@@ -10,7 +10,7 @@ contract EthereumVaultConnectorHandler is EthereumVaultConnectorHarness {
 
     function handlerCallback(
         address onBehalfOfAccount,
-        uint value,
+        uint256 value,
         bytes calldata data
     ) public payable returns (bytes memory result) {
         result = super.callback(onBehalfOfAccount, value, data);
@@ -24,10 +24,7 @@ contract CallbackTest is Test {
     EthereumVaultConnectorHandler internal evc;
 
     event CallWithContext(
-        address indexed caller,
-        address indexed targetContract,
-        address indexed onBehalfOfAccount,
-        bytes4 selector
+        address indexed caller, address indexed targetContract, address indexed onBehalfOfAccount, bytes4 selector
     );
 
     function setUp() public {
@@ -42,11 +39,7 @@ contract CallbackTest is Test {
         revert();
     }
 
-    function test_Callback(
-        address alice,
-        bytes memory data,
-        uint96 seed
-    ) public {
+    function test_Callback(address alice, bytes memory data, uint96 seed) public {
         vm.assume(alice != address(0) && alice != address(evc));
         address controller = address(new Vault(evc));
 
@@ -54,11 +47,7 @@ contract CallbackTest is Test {
         vm.deal(address(this), seed);
         vm.expectEmit(true, true, true, true, address(evc));
         emit CallWithContext(address(this), address(this), alice, bytes4(data));
-        bytes memory result = evc.handlerCallback{value: seed}(
-            alice,
-            seed,
-            data
-        );
+        bytes memory result = evc.handlerCallback{value: seed}(alice, seed, data);
         assertEq(keccak256(result), keccak256(data));
 
         // then, test with a function selector
@@ -67,25 +56,14 @@ contract CallbackTest is Test {
         evc.reset();
         Vault(controller).reset();
 
-        data = abi.encodeWithSelector(
-            Target(controller).callbackTest.selector,
-            address(evc),
-            address(evc),
-            seed,
-            alice
-        );
+        data = abi.encodeWithSelector(Target(controller).callbackTest.selector, address(evc), address(evc), seed, alice);
 
         vm.deal(controller, seed);
         vm.expectEmit(true, true, true, true, address(evc));
-        emit CallWithContext(
-            controller,
-            controller,
-            alice,
-            Target.callbackTest.selector
-        );
+        emit CallWithContext(controller, controller, alice, Target.callbackTest.selector);
         vm.prank(controller);
         result = evc.handlerCallback{value: seed}(alice, seed, data);
-        assertEq(abi.decode(result, (uint)), seed);
+        assertEq(abi.decode(result, (uint256)), seed);
     }
 
     function test_RevertIfDepthExceeded_Callback(address alice) external {
@@ -98,10 +76,7 @@ contract CallbackTest is Test {
         evc.callback(alice, 0, "");
     }
 
-    function test_RevertIfChecksReentrancy_Callback(
-        address alice,
-        uint seed
-    ) public {
+    function test_RevertIfChecksReentrancy_Callback(address alice, uint256 seed) public {
         vm.assume(alice != address(evc));
 
         evc.setChecksLock(true);
@@ -111,10 +86,7 @@ contract CallbackTest is Test {
         evc.callback{value: seed}(alice, seed, "");
     }
 
-    function test_RevertIfImpersonateReentrancy_Callback(
-        address alice,
-        uint seed
-    ) public {
+    function test_RevertIfImpersonateReentrancy_Callback(address alice, uint256 seed) public {
         vm.assume(alice != address(evc));
 
         evc.setImpersonateLock(true);
@@ -124,10 +96,7 @@ contract CallbackTest is Test {
         evc.callback{value: seed}(alice, seed, "");
     }
 
-    function test_RevertIfMsgSenderNotAuthorized_Callback(
-        address alice,
-        uint seed
-    ) public {
+    function test_RevertIfMsgSenderNotAuthorized_Callback(address alice, uint256 seed) public {
         vm.assume(alice != address(0));
         vm.assume(alice != address(evc));
 
@@ -138,10 +107,7 @@ contract CallbackTest is Test {
         evc.callback{value: seed}(alice, seed, "");
     }
 
-    function test_RevertIfValueExceedsBalance_Call(
-        address alice,
-        uint128 seed
-    ) public {
+    function test_RevertIfValueExceedsBalance_Call(address alice, uint128 seed) public {
         vm.assume(alice != address(0) && alice != address(evc));
         vm.assume(seed > 0);
 
@@ -154,15 +120,11 @@ contract CallbackTest is Test {
         evc.callback{value: seed}(alice, seed, "");
     }
 
-    function test_RevertIfInternalCallIsUnsuccessful_Callback(
-        address alice
-    ) public {
+    function test_RevertIfInternalCallIsUnsuccessful_Callback(address alice) public {
         vm.assume(alice != address(0));
         vm.assume(alice != address(evc));
 
-        bytes memory data = abi.encodeWithSelector(
-            this.revertEmptyTest.selector
-        );
+        bytes memory data = abi.encodeWithSelector(this.revertEmptyTest.selector);
 
         vm.expectRevert(Errors.EVC_EmptyError.selector);
         evc.callback(alice, 0, data);
