+++ test/unit/EthereumVaultConnector/CollateralsManagement.sol
@@ -37,22 +37,13 @@ contract EthereumVaultConnectorHandler is EthereumVaultConnectorHarness {
 contract CollateralsManagementTest is Test {
     EthereumVaultConnectorHandler internal evc;
 
-    event CollateralStatus(
-        address indexed account,
-        address indexed collateral,
-        bool enabled
-    );
+    event CollateralStatus(address indexed account, address indexed collateral, bool enabled);
 
     function setUp() public {
         evc = new EthereumVaultConnectorHandler();
     }
 
-    function test_CollateralsManagement(
-        address alice,
-        uint8 subAccountId,
-        uint8 numberOfVaults,
-        uint seed
-    ) public {
+    function test_CollateralsManagement(address alice, uint8 subAccountId, uint8 numberOfVaults, uint256 seed) public {
         // call setUp() explicitly for Dilligence Fuzzing tool to pass
         setUp();
 
@@ -67,13 +58,8 @@ contract CollateralsManagementTest is Test {
 
         // test collaterals management with use of an operator
         address msgSender = alice;
-        if (
-            seed % 2 == 0 &&
-            !evc.haveCommonOwner(account, address(uint160(seed)))
-        ) {
-            msgSender = address(
-                uint160(uint(keccak256(abi.encodePacked(seed))))
-            );
+        if (seed % 2 == 0 && !evc.haveCommonOwner(account, address(uint160(seed)))) {
+            msgSender = address(uint160(uint256(keccak256(abi.encodePacked(seed)))));
             vm.prank(alice);
             evc.setAccountOperator(account, msgSender, true);
             assertEq(evc.getAccountOwner(account), alice);
@@ -88,20 +74,15 @@ contract CollateralsManagementTest is Test {
         }
 
         // enabling collaterals
-        for (uint i = 1; i <= numberOfVaults; ++i) {
+        for (uint256 i = 1; i <= numberOfVaults; ++i) {
             Vault(controller).clearChecks();
             address[] memory collateralsPre = evc.getCollaterals(account);
 
-            address vault = i % 5 == 0
-                ? collateralsPre[seed % collateralsPre.length]
-                : address(new Vault(evc));
+            address vault = i % 5 == 0 ? collateralsPre[seed % collateralsPre.length] : address(new Vault(evc));
 
             bool alreadyEnabled = evc.isCollateralEnabled(account, vault);
 
-            assert(
-                (alreadyEnabled && i % 5 == 0) ||
-                    (!alreadyEnabled && i % 5 != 0)
-            );
+            assert((alreadyEnabled && i % 5 == 0) || (!alreadyEnabled && i % 5 != 0));
 
             if (!alreadyEnabled) {
                 vm.expectEmit(true, true, false, true, address(evc));
@@ -119,7 +100,7 @@ contract CollateralsManagementTest is Test {
                 assertEq(collateralsPost[collateralsPost.length - 1], vault);
             }
 
-            for (uint j = 0; j < collateralsPre.length; ++j) {
+            for (uint256 j = 0; j < collateralsPre.length; ++j) {
                 assertEq(collateralsPre[j], collateralsPost[j]);
             }
         }
@@ -139,22 +120,14 @@ contract CollateralsManagementTest is Test {
 
             assertEq(collateralsPost.length, collateralsPre.length - 1);
 
-            for (uint j = 0; j < collateralsPost.length; ++j) {
+            for (uint256 j = 0; j < collateralsPost.length; ++j) {
                 assertNotEq(collateralsPost[j], vault);
             }
         }
     }
 
-    function test_RevertIfNotOwnerOrNotOperator_CollateralsManagement(
-        address alice,
-        address bob
-    ) public {
-        vm.assume(
-            alice != address(0) &&
-                alice != address(evc) &&
-                bob != address(0) &&
-                bob != address(evc)
-        );
+    function test_RevertIfNotOwnerOrNotOperator_CollateralsManagement(address alice, address bob) public {
+        vm.assume(alice != address(0) && alice != address(evc) && bob != address(0) && bob != address(evc));
         vm.assume(!evc.haveCommonOwner(alice, bob));
 
         address vault = address(new Vault(evc));
@@ -177,9 +150,7 @@ contract CollateralsManagementTest is Test {
         evc.disableCollateral(bob, vault);
     }
 
-    function test_RevertIfChecksReentrancy_CollateralsManagement(
-        address alice
-    ) public {
+    function test_RevertIfChecksReentrancy_CollateralsManagement(address alice) public {
         vm.assume(alice != address(evc));
         address vault = address(new Vault(evc));
 
@@ -206,9 +177,7 @@ contract CollateralsManagementTest is Test {
         evc.disableCollateral(alice, vault);
     }
 
-    function test_RevertIfImpersonateReentrancy_CollateralsManagement(
-        address alice
-    ) public {
+    function test_RevertIfImpersonateReentrancy_CollateralsManagement(address alice) public {
         vm.assume(alice != address(evc));
         address vault = address(new Vault(evc));
 
@@ -235,18 +204,14 @@ contract CollateralsManagementTest is Test {
         evc.disableCollateral(alice, vault);
     }
 
-    function test_RevertIfInvalidVault_CollateralsManagement(
-        address alice
-    ) public {
+    function test_RevertIfInvalidVault_CollateralsManagement(address alice) public {
         vm.assume(alice != address(evc));
         vm.prank(alice);
         vm.expectRevert(Errors.EVC_InvalidAddress.selector);
         evc.enableCollateral(alice, address(evc));
     }
 
-    function test_RevertIfAccountStatusViolated_CollateralsManagement(
-        address alice
-    ) public {
+    function test_RevertIfAccountStatusViolated_CollateralsManagement(address alice) public {
         vm.assume(alice != address(evc));
 
         address vault = address(new Vault(evc));
